+++ Content.Shared/_CP14/MagicSpell/CP14SharedMagicSystem.cs
@@ -1,8 +1,8 @@
-using Content.Shared._CP14.Magic.Components;
-using Content.Shared._CP14.Magic.Events;
 using Content.Shared._CP14.MagicEnergy;
 using Content.Shared._CP14.MagicEnergy.Components;
-using Content.Shared.Actions;
+using Content.Shared._CP14.MagicSpell.Components;
+using Content.Shared._CP14.MagicSpell.Events;
+using Content.Shared._CP14.MagicSpell.Interfaces;
 using Content.Shared.DoAfter;
 using Content.Shared.Hands.Components;
 using Content.Shared.Popups;
@@ -13,7 +13,7 @@ using Robust.Shared.Network;
 using Robust.Shared.Physics.Systems;
 using Robust.Shared.Random;
 
-namespace Content.Shared._CP14.Magic;
+namespace Content.Shared._CP14.MagicSpell;
 
 /// <summary>
 /// This system handles the basic mechanics of spell use, such as doAfter, event invocation, and energy spending.
@@ -29,7 +29,6 @@ public partial class CP14SharedMagicSystem : EntitySystem
     [Dependency] private readonly SharedCP14MagicEnergySystem _magicEnergy = default!;
     [Dependency] private readonly SharedPopupSystem _popup = default!;
     [Dependency] private readonly IRobustRandom _random = default!;
-    [Dependency] private readonly ActionContainerSystem _actionContainer = default!;
 
     public override void Initialize()
     {
@@ -49,7 +48,6 @@ public partial class CP14SharedMagicSystem : EntitySystem
         SubscribeLocalEvent<CP14DelayedWorldTargetActionEvent>(OnWorldTargetAction);
 
         InitializeSpells();
-        InitializeSpellStorage();
     }
 
     private void OnSomaticAspectBeforeCast(Entity<CP14MagicEffectSomaticAspectComponent> ent, ref CP14BeforeCastMagicEffectEvent args)
