+++ Content.Shared/_CP14/MagicSpell/Components/CP14MagicEffectVerbalAspectComponent.cs
@@ -1,6 +1,4 @@
-using Content.Shared.FixedPoint;
-
-namespace Content.Shared._CP14.Magic.Components;
+namespace Content.Shared._CP14.MagicSpell.Components;
 
 /// <summary>
 /// Requires the user to be able to speak in order to use this spell. Also forces the user to use certain phrases at the beginning and end of a spell cast
