+++ Content.Shared/_CP14/MagicSpell/Components/Spells/CP14DelayedApplyEntityEffectsSpellComponent.cs
@@ -1,12 +1,12 @@
 using Content.Shared.EntityEffects;
 
-namespace Content.Shared._CP14.Magic.Components.Spells;
+namespace Content.Shared._CP14.MagicSpell.Components.Spells;
 
 /// <summary>
 /// Stores a list of effects for delayed actions.
 /// </summary>
 [RegisterComponent, Access(typeof(CP14SharedMagicSystem))]
-public sealed partial class CP14DelayedSelfEntityEffectSpellComponent : Component
+public sealed partial class CP14DelayedApplyEntityEffectsSpellComponent : Component
 {
     [DataField(required: true, serverOnly: true)]
     public List<EntityEffect> Effects = new();
