+++ Content.Shared/_CP14/MagicSpell/Components/Spells/CP14DelayedSpawnEntitiesSpellComponent.cs
@@ -1,12 +1,12 @@
 using Robust.Shared.Prototypes;
 
-namespace Content.Shared._CP14.Magic.Components.Spells;
+namespace Content.Shared._CP14.MagicSpell.Components.Spells;
 
 /// <summary>
 /// Stores a list of effects for delayed actions.
 /// </summary>
 [RegisterComponent, Access(typeof(CP14SharedMagicSystem))]
-public sealed partial class CP14DelayedSpawnOnWorldTargetSpellComponent : Component
+public sealed partial class CP14DelayedSpawnEntitiesSpellComponent : Component
 {
     /// <summary>
     /// What entities should be spawned.
